requirements.txt
beautifulsoup4
requests
pinecone-client>=3.0.0
python-dotenv
sentence-transformers>=2.2.0
